Refactored the conflict module
