+++ src/civilization/conflict/components.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::{Component, Reflect};
+
+#[derive(Component, Debug, Reflect)]
+pub struct UnresolvedConflict;
