+++ src/civilization/conflict/mod.rs
@@ -1,1 +1,3 @@
 pub mod plugin;
+mod systems;
+mod components;
